Targeted refactor on invoice endpoints, MCP contracts and DB seeders
